requirements.txt
azure-ai-agents==1.2.0b4
azure-ai-projects==1.1.0b4
azure-core==1.35.1
azure-identity==1.25.0
azure-storage-blob==12.26.0
semantic-kernel==1.37.0
